HIVE-27180: Remove JsonSerde from hcatalog, Upgrade should update changed FQN for JsonSerDe in HMS DB
Fixing imports
