+++ hcatalog/src/test/e2e/hcatalog/tests/hadoop.conf
@@ -163,26 +163,6 @@ jar :FUNCPATH:/testudf.jar org.apache.hive.hcatalog.utils.WriteText -libjars :HC
                                 ,'delimiter' => '	'
                                 },
                                 {
-                                 'num' => 2
-                                ,'hcat_prep' => q\
-drop table if exists hadoop_write_2;
-create table hadoop_write_2(
-            s string,
-            i int,
-            d double,
-            m map<string, string>,
-            bb array<struct<a: int, b: string>>)
-            row format serde 'org.apache.hive.hcatalog.data.JsonSerDe'
-            stored as textfile;\
-                                ,'hadoop' => q\
-jar :FUNCPATH:/testudf.jar org.apache.hive.hcatalog.utils.WriteJson -libjars :HCAT_JAR: :THRIFTSERVER: all100kjson hadoop_write_2
-\,
-                                ,'result_table' => 'hadoop_write_2'
-                                ,'sql' => q\select s, i, d, '', '' from all100kjson;\
-                                ,'floatpostprocess' => 1
-                                ,'delimiter' => '	'
-                                },
-                                {
                                  'num' => 3
                                 ,'hcat_prep' => q\
 drop table if exists hadoop_write_3;
