+++ hcatalog/src/test/e2e/hcatalog/tests/hive.conf
@@ -164,16 +164,6 @@ insert into TABLE hive_write_1 select t, si, i, b, s from all100k;\,
                          'floatpostprocess' => 1,
                          'delimiter' => '	',
                     },{
-                        'num' => 2,
-                         'sql' => q\
-drop table if exists hive_write_2;
-create table hive_write_2 (name string, age int, gpa double) row format serde 'org.apache.hive.hcatalog.data.JsonSerDe' stored as textfile;
-insert into TABLE hive_write_2 select s, i, 0.1 from all100kjson;\,
-                         'result_table' => 'hive_write_2',
-                         'verify_sql' =>"select s, i, 0.1 from all100kjson;",
-                         'floatpostprocess' => 1,
-                         'delimiter' => '	',
-                    },{
                         'num' => 3,
                          'sql' => q\
 drop table if exists hive_write_3;
