+++ hcatalog/src/test/e2e/hcatalog/tests/pig.conf
@@ -203,26 +203,6 @@ store a into 'pig_write_1' using org.apache.hive.hcatalog.pig.HCatStorer();\,
                                 ,'delimiter' => '	'
                                 },
                                 {
-                                 'num' => 2
-                                ,'hcat_prep'=>q\drop table if exists pig_write_2;
-create table pig_write_2(
-            s string,
-            i int,
-            d double,
-            m map<string, string>,
-            bb array<struct<a: int, b: string>>)
-            row format serde 'org.apache.hive.hcatalog.data.JsonSerDe'
-            STORED AS TEXTFILE;
-\
-                                ,'pig' => q\a = load 'all100kjson' using org.apache.hive.hcatalog.pig.HCatLoader();
-b = foreach a generate s, i, d;
-store b into 'pig_write_2' using org.apache.hive.hcatalog.pig.HCatStorer();\,
-				,'sql' => q\select IFNULL(s, ""), IFNULL(i, ""), IFNULL(d, "") from all100kjson;\
-                                ,'result_table' => 'pig_write_2'
-                                ,'floatpostprocess' => 1
-                                ,'delimiter' => '	'
-                                },
-                                {
                                  'num' => 3
                                 ,'hcat_prep'=>q\drop table if exists pig_write_3;
 create table pig_write_3(
