+++ ql/src/test/queries/clientpositive/json_serde_tsformat.q
@@ -1,8 +1,6 @@
 
-add jar ${system:maven.local.repository}/org/apache/hive/hcatalog/hive-hcatalog-core/${system:hive.version}/hive-hcatalog-core-${system:hive.version}.jar;
-
 CREATE TABLE t1_n156 (c1 int, c2 string, c3 timestamp)
-ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
+ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.JsonSerDe'
 WITH SERDEPROPERTIES ('timestamp.formats'='yyyy-MM-dd\'T\'HH:mm:ss')
 ;
 LOAD DATA LOCAL INPATH "../../data/files/tsformat.json" INTO TABLE t1_n156;
