+++ ql/src/test/queries/clientpositive/parquet_mixed_partition_formats2.q
@@ -9,7 +9,7 @@ reports array<bigint> COMMENT 'from deserializer')
 PARTITIONED BY (
 ts string)
 ROW FORMAT SERDE
-'org.apache.hive.hcatalog.data.JsonSerDe'
+'org.apache.hadoop.hive.serde2.JsonSerDe'
 STORED AS INPUTFORMAT
 'org.apache.hadoop.mapred.TextInputFormat'
 OUTPUTFORMAT
