+++ ql/src/test/results/clientpositive/llap/json_serde_qualified_types.q.out
@@ -2,7 +2,7 @@ PREHOOK: query: create table json_serde_qualified_types (
   c1 char(10),
   c2 varchar(20),
   c3 decimal(10, 5)
-) row format serde 'org.apache.hive.hcatalog.data.JsonSerDe'
+) row format serde 'org.apache.hadoop.hive.serde2.JsonSerDe'
 PREHOOK: type: CREATETABLE
 PREHOOK: Output: database:default
 PREHOOK: Output: default@json_serde_qualified_types
@@ -10,7 +10,7 @@ POSTHOOK: query: create table json_serde_qualified_types (
   c1 char(10),
   c2 varchar(20),
   c3 decimal(10, 5)
-) row format serde 'org.apache.hive.hcatalog.data.JsonSerDe'
+) row format serde 'org.apache.hadoop.hive.serde2.JsonSerDe'
 POSTHOOK: type: CREATETABLE
 POSTHOOK: Output: database:default
 POSTHOOK: Output: default@json_serde_qualified_types
