+++ ql/src/test/results/clientpositive/llap/json_serde_tsformat.q.out
@@ -1,11 +1,11 @@
 PREHOOK: query: CREATE TABLE t1_n156 (c1 int, c2 string, c3 timestamp)
-ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
+ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.JsonSerDe'
 WITH SERDEPROPERTIES ('timestamp.formats'='yyyy-MM-dd\'T\'HH:mm:ss')
 PREHOOK: type: CREATETABLE
 PREHOOK: Output: database:default
 PREHOOK: Output: default@t1_n156
 POSTHOOK: query: CREATE TABLE t1_n156 (c1 int, c2 string, c3 timestamp)
-ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
+ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.JsonSerDe'
 WITH SERDEPROPERTIES ('timestamp.formats'='yyyy-MM-dd\'T\'HH:mm:ss')
 POSTHOOK: type: CREATETABLE
 POSTHOOK: Output: database:default
