+++ ql/src/test/results/clientpositive/llap/parquet_mixed_partition_formats2.q.out
@@ -5,7 +5,7 @@ reports array<bigint> COMMENT 'from deserializer')
 PARTITIONED BY (
 ts string)
 ROW FORMAT SERDE
-'org.apache.hive.hcatalog.data.JsonSerDe'
+'org.apache.hadoop.hive.serde2.JsonSerDe'
 STORED AS INPUTFORMAT
 'org.apache.hadoop.mapred.TextInputFormat'
 OUTPUTFORMAT
@@ -20,7 +20,7 @@ reports array<bigint> COMMENT 'from deserializer')
 PARTITIONED BY (
 ts string)
 ROW FORMAT SERDE
-'org.apache.hive.hcatalog.data.JsonSerDe'
+'org.apache.hadoop.hive.serde2.JsonSerDe'
 STORED AS INPUTFORMAT
 'org.apache.hadoop.mapred.TextInputFormat'
 OUTPUTFORMAT
