+++ ql/src/test/results/clientpositive/llap/schema_evol_orc_acid_part_update.q.out
@@ -618,7 +618,7 @@ PREHOOK: query: CREATE TABLE all100kjson_textfile_orc_n0 (
                              bo boolean,
                              ts timestamp)
                              PARTITIONED BY (t tinyint)
-                             ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
+                             ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.JsonSerDe'
                              WITH SERDEPROPERTIES ('timestamp.formats'='yyyy-MM-dd\'T\'HH:mm:ss')
                              STORED AS TEXTFILE
 PREHOOK: type: CREATETABLE
@@ -634,7 +634,7 @@ POSTHOOK: query: CREATE TABLE all100kjson_textfile_orc_n0 (
                              bo boolean,
                              ts timestamp)
                              PARTITIONED BY (t tinyint)
-                             ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
+                             ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.JsonSerDe'
                              WITH SERDEPROPERTIES ('timestamp.formats'='yyyy-MM-dd\'T\'HH:mm:ss')
                              STORED AS TEXTFILE
 POSTHOOK: type: CREATETABLE
