+++ standalone-metastore/metastore-server/src/main/sql/oracle/upgrade-4.0.0-alpha-2-to-4.0.0.oracle.sql
@@ -20,6 +20,9 @@ CREATE TABLE MIN_HISTORY_WRITE_ID (
   MH_WRITEID NUMBER(19) NOT NULL
 );
 
+-- HIVE-27180
+UPDATE SERDES SET SLIB='org.apache.hadoop.hive.serde2.JsonSerDe' where SLIB='org.apache.hive.hcatalog.data.JsonSerDe';
+
 -- These lines need to be last.  Insert any changes above.
 UPDATE VERSION SET SCHEMA_VERSION='4.0.0', VERSION_COMMENT='Hive release version 4.0.0' where VER_ID=1;
 SELECT 'Finished upgrading MetaStore schema from 4.0.0-alpha-2 to 4.0.0' AS Status from dual;
